+++ src/main/java/com/alexsid/lesson17/DuplicatedPetExceptionHandler.java
@@ -1,6 +1,6 @@
-package com.alexsid.lesson1517;
+package com.alexsid.lesson17;
 
-import com.alexsid.lesson1517.pets.DuplicatedPetException;
+import com.alexsid.lesson17.pets.DuplicatedPetException;
 
 public interface DuplicatedPetExceptionHandler {//strategy pattern
     void process(DuplicatedPetException e);
