+++ src/main/java/com/alexsid/lesson17/Main.java
@@ -1,4 +1,4 @@
-package com.alexsid.lesson1517;
+package com.alexsid.lesson17;
 
 public class Main {
     public static void main(String[] args) {
