+++ src/main/java/com/alexsid/lesson17/pets/Dog.java
@@ -1,13 +1,13 @@
-package com.alexsid.lesson1517.pets;
+package com.alexsid.lesson17.pets;
 
-import com.alexsid.lesson1517.Person;
+import com.alexsid.lesson17.Person;
 
 import java.util.Objects;
 
-public class Cat extends AbstractPet {
-    private final PetType type = PetType.CAT;
+public class Dog extends AbstractPet {
+    private final PetType type = PetType.DOG;
 
-    protected Cat(int id, String name, Person owner, double weight) {
+    protected Dog(int id, String name, Person owner, double weight) {
         super(id, name, owner, weight);
     }
 
@@ -19,7 +19,7 @@ public class Cat extends AbstractPet {
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
-        if (!(o instanceof Cat)) return false;
+        if (!(o instanceof Dog)) return false;
         return super.equals(o);
     }
