+++ src/main/java/com/alexsid/lesson17/pets/DuplicatedPetException.java
@@ -1,4 +1,4 @@
-package com.alexsid.lesson1517.pets;
+package com.alexsid.lesson17.pets;
 
 public class DuplicatedPetException extends RuntimeException{
     {
