+++ src/main/java/com/alexsid/lesson17/pets/PetFactory.java
@@ -1,6 +1,6 @@
-package com.alexsid.lesson1517.pets;
+package com.alexsid.lesson17.pets;
 
-import com.alexsid.lesson1517.Person;
+import com.alexsid.lesson17.Person;
 
 public class PetFactory {
