+++ src/main/java/com/alexsid/lesson17/pets/PetType.java
@@ -1,4 +1,4 @@
-package com.alexsid.lesson1517.pets;
+package com.alexsid.lesson17.pets;
 
 public enum PetType {
     CAT,
